requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
editdistance==0.8.1
filelock==3.18.0
fonttools==4.58.2
frozenlist==1.6.2
fsspec==2025.3.0
h5py==3.14.0
huggingface-hub==0.32.4
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.1
pyarrow==20.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sympy==1.14.0
torch==2.7.1+cu118
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
